remove child as selection for a parent model (#2822)
Will close: https://github.com/zesty-io/manager-ui/issues/2580

As a solution for the issue above. I make sure that when an item is a
child it should not be in a list where we can set a model parent.

To show how it resolves the issue. I set the parent of testDataSet to
testPage, then go to testPage to try check if testDataSet is present as
a model parent in the list which is not.


[screencast-8-9099fdbffb-hsbtls.manager.dev.zesty.io_8080-2024.07.17-20_39_01.webm](https://github.com/user-attachments/assets/3d179fd3-484f-456e-a9a9-5c7798e6b1e9)