[#3335] Blocks: Starter Kit - Bugs and Enhancements (#3343)
### Bugs
- [x] **Error flow for model creation**: The current error flow when a
model name and reference ID already exist is missing, causing a redirect
to the "Select Starter Kit" page. This is not the desired behavior. This
does not happen each time though.
- [x] **Redirect after Block creation**: After the user creates a Block
with a starter kit, the user should be redirected to the newly created
Block in the Blocks App.
- [x] **Modal height** should remain fixed when the user performs a
search.

### Enhancements
- [x] **Blank Model Modal**: The blank model should use the classic
"Create Block Model" modal.
- [x] Change "Done" button text to **"Create Block Model"** 
- [x]  Hover and selected states for the card.


https://github.com/user-attachments/assets/cd3778f9-0dd7-4fc7-a218-06f622f24d91