Blocks-App-and-Modals-Enhancements (#3370)
https://github.com/zesty-io/manager-ui/pull/3345#issuecomment-2764476808

## Modals

- [x] **Backdrop:** Implement the new Design System backdrop for all
modals across the content, schema, and media apps.

- [x] **Modal Consistency:** Resolve the following interlinked issues:
- [x] **Modal: Icon Size and Padding:** Standardize icon sizes, padding,
and spacing across all modals.
    
# Blocks App

- [x] Create Block Modal Height is not set correctly. Please ensure the
Modal height occupies the viewport's height with 20px padding at the top
and 20px padding at the bottom. Add a minimum height of 680px and
maximum height of 1240px. Please see how this same modal height behavior
was implemented for the Select Item dialog modal of the One to Many
field as a reference


## Enhancements

- [x] Create Block and Create Model Pop Up: Select First Option in list
by default