[Chore] Resolve failing Redirects test (#3781)
Resolve an issue where the apiRef is just an empty object due to a race
condition. Due to apiRef not being able to trigger a useEffect, the fix
is to add a boolean flag that will re-trigger the
`RedirectsDelete.tsx`'s useEffect whenever the grid table has loaded.
This ensures that when the useEffect is re-triggered, the apiRef will
already be initialized and we can subscribe to any events that it has.